framework: guard latent integer-cast wraps in time/limit math
Sweep for the same truncation class as the payment-quantity fix: rate_limit
Bucket::retry_after cast usize->u32 (asymmetric with its try_record sibling);
workflow lock-timeout, remember-me lifetime, and queue visibility/requeue
durations cast u64->i64 uncapped, wrapping to a negative (past) timestamp on
extreme config and thrashing the affected store. Use checked/saturating casts
and add an upper-bound check to WorkflowConfig::validate.

+++ framework/src/rate_limit/algorithm.rs
@@ -46,7 +46,7 @@ impl Bucket {
     /// when the bucket is under the limit and a new hit would succeed
     /// immediately.
     pub fn retry_after(&self, max: u32, window: Duration, now: Instant) -> Option<Duration> {
-        if (self.hits.len() as u32) < max {
+        if self.hits.len() < max as usize {
             return None;
         }
         let oldest = self.hits.front().copied()?;
